Bugfix with file reads - read description
Fix a bug where the program went over the entire buffer each time it read the file, even if fread didn't fill the buffer. I changed the end condition for the for loop to be the number of chars read, and made sure that the entire buffer is cleared each time.